requirements.txt
fastapi
uvicorn
pydantic
RGBMatrixEmulator
Pillow
